requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
ffmpeg-python==0.2.0
fonttools==4.39.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
joblib==1.2.0
keyboard==0.13.5
kiwisolver==1.4.4
lazy_loader==0.1
librosa==0.10.0
llvmlite==0.39.1
markdown-it-py==2.2.0
matplotlib==3.7.1
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
noisereduce==2.0.1
numba==0.56.4
numpy==1.23.5
openai==0.27.1
packaging==23.0
Pillow==9.4.0
platformdirs==3.1.1
pooch==1.7.0
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.1
pyttsx3==2.90
requests==2.28.2
rfc3986==1.5.0
rich==13.3.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.4
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
